pyproject.toml
[project]
name = "rag-document-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "watchdog>=6.0.0",
]
